requirements.txt
openai-whisper
python-dotenv
numpy
torch
ffmpeg-python
